Edge Behavior
Hopefully this thread will be able to sense whether there is an edge or
not. It does this by its "edge" variable. If it doesn't have any red
light reflecting back to it "edge" is true and the robot should try to
adjust by shutting off left wheel. This turn mechanic will probably need
to be changed or adjusted after testing.

Also, this is using the motors how the original behavior example uses
them because the differential motor thread was not created yet.